refactor(js): Extract guards and add computed property access error for stdlib members
- Refactored executeStdlibMemberExpression to use consistent guard pattern
- Extracted guards: guardPropertyIsNotStub, guardPropertyIsAllowed, guardMethodIsAllowed
- Added handleStdlibError for consistent error handling
- Added proper TypeError for computed access on stdlib members (arr['length'] now throws error)
- Added tests for computed property access errors on arrays
- Fixed ESLint warnings about type narrowing
- All existing tests remain green
