+++ src/javascript/describers/describeMemberExpression.ts
@@ -5,9 +5,9 @@ export function describeMemberExpression(
   expression: MemberExpression,
   result: EvaluationResultMemberExpression
 ): string {
-  const jikiObject = result.immutableJikiObject || result.jikiObject;
-  const _objectValue = result.object.immutableJikiObject || result.object.jikiObject;
-  const indexValue = result.property.immutableJikiObject || result.property.jikiObject;
+  const jikiObject = result.immutableJikiObject;
+  const _objectValue = result.object.immutableJikiObject;
+  const indexValue = result.property.immutableJikiObject;
 
   return `Accessed element at index ${indexValue.toString()} of the list, got ${jikiObject.toString()}`;
 }
